requirements.txt
tensorflow_gpu==2.3.1
opencv_python==4.4.0.42
scikit_image==0.17.2
numpy==1.18.5
PyYAML==5.3.1
